+++ src/heap.c
@@ -16,6 +16,7 @@
 */
 
 #include <stdio.h>
+#include <stdlib.h>
 #include <stdint.h>
 #include <inttypes.h>
 
@@ -92,7 +93,7 @@ int generate_freelist(size_t num_cells) {
   return 1;
 }
 
-int heap_init(size_t num_cells) {
+int heap_init(int num_cells) {
 
   // retrieve nil symbol value f
   SYMBOL_NIL = symrepr_nil();
